requirements.txt
langchain
langchain_chroma
langchain-openai
langchain-core
streamlit
python-dotenv
pandas